added http interception and changed login url configuration to own provider
Needed to change how one can configure the login url since the HttpInterceptor should not broadcast an event in case a login attempt was made. Therefore, I introduced an own factory for the login url which simply can be redefined if necessary and is used by the http interceptor as well as the AuthService. Modified the README.md accordingly.